Write and read comments into and out of the db
This is a pretty big changeset:

* Added dependencies for slick, joda-time, and some others
* Restructured startup code so that the API doesn't become
  available until migrations have been completed. This was
  a huge change that basically converted the main class into
  an Actor, and now that Actor handles orchestrating startup.
* And, obviously, fleshed out the TipsApi so that now you can
  post a tip and then pull it back. A lot of other work was
  needed to support that feature, and the restructuring of
  startup was required to make it work consistently; prior to
  reworking startup, there was a race between when the database
  connection pool was filled and when the migrations had completed.
  Any connections made prior to completing the migrations would
  be set to the wrong search_path, and unable to find the `tips`
  table.

+++ src/it/scala/matchers/TipsMatchers.scala
@@ -20,7 +20,7 @@ package tipster.test.matchers
 
 import org.scalatest.matchers._
 
-import tipster.tips._
+import tipster.tips.model._
 
 trait TipsMatchers {
   def username[A <: HasUsername](expected: String) = new HavePropertyMatcher[A, String] {
@@ -43,7 +43,7 @@ trait TipsMatchers {
       )
   }
 
-  def id[A <: HasId](expected: Long) = new HavePropertyMatcher[A, Long] {
+  def id[A <: HasId](expected: Int) = new HavePropertyMatcher[A, Int] {
     def apply(message: A) =
       HavePropertyMatchResult(
         message.id == expected,
